Moved out TurnEngineEnums
